requirements.txt
openai>=1.0.0
pyyaml>=6.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
celery>=5.3.0
redis>=5.0.0
pydantic-settings>=2.0.0
# FastAPI 后端依赖
fastapi>=0.100.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
# 认证
passlib>=1.7
bcrypt>=4.0
python-jose[cryptography]>=3.5
cryptography>=41.0
python-multipart>=0.0.6
# 导出功能
ebooklib>=0.18
python-docx>=1.1